2024-2/Sanhak Project - Sleep Management Service/SleepDiver/requirements.txt
asgiref==3.8.1
Django==5.1.1
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
wheel==0.44.0
gunicorn==20.1.0
django-cors-headers
djangorestframework>=3.15.1
drf-yasg>=1.21.8
python-decouple 
openai
python-dotenv
mysqlclient
cryptography